Significant changes see details
Added data structure:
1) to hold logon credendials (map mLogOnCredentials Map<String,
String>()
2) to hold updated log strings FixedStringLog. (this is going to hold
info of what's happening to twitter). And size constant.

Also fixed constructor of TweetDirector, so that it throws exception in
case null credentials supplied.

Made a stub for tweetTheMessage() method, now we kinda can run this
class to test. TODO: actual tweeting functionality.

Added shutdown() method contents.

Fixed getLogObject() method thought which we can figure out what's
happening with the thing.
Should this object be actually synchronized??? (it is going to be
between TweetDirector and TweetThead)

Added getLogObject() method (so that responsible for the display of the
tweet object can access status).